Update dagger tests
